+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # 2.x branch
 ## 2.31 branch
 ### 2.31.4
+* `login` layout changed to `single-column`. This name clarifies that it can be used for various purposes (login, sign
+  up, but also error views);
 * removed the `cookies_policy` template element, the `SystemsController::acceptCookies()` method and all code related to
   privacy for cookies (you need to integrate your privacy cookie);
 * added `MeCms.default.other_css` and `MeCms.default.other_js` configuration values. These are used directly by the